Move the ticker to the exp directory (#1101)
* Move the ticker to the exp directory

* Fix import
