+++ exp/ticker/.gitignore
