+++ exp/ticker/docs/images/StellarTicker.png
