+++ exp/ticker/internal/scraper/asset_scraper.go
