+++ exp/ticker/internal/scraper/asset_scraper_test.go
