+++ exp/ticker/internal/scraper/main.go
